more lazyLoading and standalone + new routing init
