+++ src/app/_components/_forms/save-city-form/save-city-form.component.html
@@ -1,4 +1,4 @@
-<script src="../../../home/home.component.ts"></script>
+<script src="../../home/home.component.ts"></script>
 <script src="../../../_service/api-data.service.ts"></script>
 <form [formGroup]="searchCityForm" (ngSubmit)="saveNewCity()">
     <mat-form-field>
