+++ src/app/_components/home/info-small/info-small.component.scss
