+++ src/app/_components/home/infoscreen/current-highlights/current-highlights.component.scss
