+++ src/app/_components/home/infoscreen/current-today/current-today.component.html
@@ -7,10 +7,10 @@
         </div>
         <hr>
         <div>
-            <img ngSrc="./assets/images/locationPin.png" [width]="24" [height]="24"> <p>{{ locationName }}</p>
+            <img ngSrc="../../../../../assets/images/locationPin.png" [width]="24" [height]="24"> <p>{{ locationName }}</p>
         </div>
         <div>
-            <img ngSrc="./assets/images/date.png" [width]="24" [height]="24"> <p>{{ date }}</p>
+            <img ngSrc="../../../../../assets/images/date.png" [width]="24" [height]="24"> <p>{{ date }}</p>
         </div>
     </div>
 </main>
