+++ src/app/_components/home/infoscreen/current-today/current-today.component.scss
