+++ src/app/_components/home/infoscreen/forecast/forecast.component.scss
